feat[8.23]: add subscription for adding books to backend
