+++ part8/backend/apollo/typeDefs.js
@@ -1,0 +1,60 @@
+const typeDefs = `
+  type Author {
+    name: String!
+    id: ID!
+    born: Int
+    bookCount: Int
+  }
+
+  type Book {
+    title: String!
+    published: Int!
+    author: Author!
+    id: ID!
+    genres: [String!]
+  }
+
+  type User {
+    username: String!
+    favoriteGenre: String!
+    id: ID!
+  }
+
+  type Token {
+    value: String!
+  }
+
+  type Query {
+    allBooks(author: String, genre: String): [Book]
+    favoriteBooks: [Book]
+    bookCount: Int
+    allAuthors: [Author]
+    allUsers: [User]
+    authorCount: Int
+    me: User
+  }
+
+  type Mutation {
+    addBook(
+      title: String!
+      author: String!
+      published: Int!
+      genres: [String!]!
+    ): Book
+    editAuthor(name: String!, setBornTo: Int!): Author
+    createUser(
+      username: String!
+      favoriteGenre: String!
+    ): User
+    login(
+      username: String!
+      password: String!
+    ): Token
+  }
+
+  type Subscription {
+    bookAdded: Book!
+  }
+`
+
+module.exports = typeDefs
